+++ R/create_slope_cs.R
@@ -6,7 +6,11 @@
 #'
 #' @details
 #'
-#' The most widely used cost function that approximates the difficulty of moving across a landscape is Waldo Tobler's 'Hiking Function' (Gorenflo and Gale, 1990; Wheatley and Gillings, 2001). This function allows for the assessment of time necessary to traverse a surface and takes into account up-slope and down-slope momentum (Kantner, 2004; Tobler, 1993).
+#' Tobler's 'Hiking Function' is the most widely used cost function when approximating the difficulty of moving across a landscape (Gorenflo and Gale, 1990; Wheatley and Gillings, 2001). The function assess the time necessary to traverse a surface and takes into account up-slope and down-slope (Kantner, 2004; Tobler, 1993).
+#'
+#' Tobler's offpath Hiking Function reduces the speed of the Tobler's Hiking Function by 0.6 to take into account walking off-path (Tobler, 1993)
+#'
+#'The Irmischer and Clark functions were modelled from speed estimates of United States Military Academy (USMA) cadets while they navigated on foot over hilly, wooded terrain as part of their summer training in map and compass navigation.
 #'
 #' The Modified Hiking cost function combines MIDE (París Roche, 2002), a method to calculate walking hours for an average hiker with a light load (Márquez-Pérez et al., 2017), and Tobler's 'Hiking Function' (Tobler, 1993). The Modified Hiking Function benefits from the precision of the MIDE rule and the continuity of Tobler's Hiking Function (Márquez-Pérez et al., 2017).
 #'
@@ -14,9 +18,11 @@
 #'
 #' Herzog's (2010) Sixth-degree polynomial cost function approximates the energy expenditure values found in Minetti et al. (2002) but eliminates the problem of unrealistic negative energy expenditure values for steep downhill slopes.
 #'
+#' Llobera and Sluckin (2007) cost function approximates the metabolic energy expenditure in KJ/(m*kg) when moving across a landscape.
+#'
 #' @param dem \code{RasterLayer} (raster package). Digital Elevation Model.
 #'
-#' @param cost_function Cost Function used in the Least Cost Path calculation. Implemented cost functions include 'tobler', 'modified tobler', 'wheeled transport' 'herzog', or 'all'. Default is 'tobler'. See Details for more information.
+#' @param cost_function Cost Function used in the Least Cost Path calculation. Implemented cost functions include 'tobler', 'tobler offpath', 'irmischer-clarke male', 'irmischer-clarke offpath male', 'irmischer-clarke female', 'irmischer-clarke offpath female', 'modified tobler', 'wheeled transport', 'herzog', 'llobera-sluckin', 'all'. Default is 'tobler'. See Details for more information.
 #'
 #' @param neighbours \code{numeric} value. Number of directions used in the Least Cost Path calculation. See Huber and Church (1985) for methodological considerations when choosing number of neighbours. Expected values are 4, 8, or 16. Default is 16.
 #'
@@ -24,7 +30,7 @@
 #'
 #' @param max_slope \code{numeric} value. Maximum percentage slope that is traversable. Default is NULL.
 #'
-#' @return \code{TransitionLayer} (gdistance package) numerically expressing the difficulty of moving up/down slope based on the cost function provided in the cost_function argument.
+#' @return \code{TransitionLayer} (gdistance package) numerically expressing the difficulty of moving up/down slope based on the cost function provided in the cost_function argument. list of \code{TransitionLayer} if cost_function = 'all'.
 #'
 #' @author Joseph Lewis
 #'
@@ -47,9 +53,11 @@ create_slope_cs <- function(dem, cost_function = "tobler", neighbours = 16, crit
         stop("dem argument is invalid. Expecting a RasterLayer object")
     }
     
-    if (cost_function != "tobler" & cost_function != "modified tobler" & cost_function != "wheeled transport" & cost_function != "herzog" & cost_function != 
-        "all") {
-        stop("cost_function argument is invalid. Expecting 'tobler', 'modified tobler', 'wheeled transport', 'herzog', or 'all'")
+    cfs <- c("tobler", "tobler offpath", "irmischer-clarke male", "irmischer-clarke offpath male", "irmischer-clarke female", "irmischer-clarke offpath female", 
+        "modified tobler", "wheeled transport", "herzog", "llobera-sluckin", "all")
+    
+    if (!cost_function %in% cfs) {
+        stop("cost_function argument is invalid. See details for accepted cost functions")
     }
     
     if (!neighbours %in% c(4, 8, 16)) {
@@ -60,7 +68,7 @@ create_slope_cs <- function(dem, cost_function = "tobler", neighbours = 16, crit
         x[2] - x[1]
     }
     
-    hd <- gdistance::transition(dem, altDiff_slope, 8, symm = FALSE)
+    hd <- suppressWarnings(gdistance::transition(dem, altDiff_slope, 8, symm = FALSE))
     
     slope <- gdistance::geoCorrection(hd)
     
@@ -79,62 +87,170 @@ create_slope_cs <- function(dem, cost_function = "tobler", neighbours = 16, crit
     }
     
     if (cost_function == "tobler") {
-        slope[adj] <- 6 * exp(-3.5 * abs(slope[adj] + 0.05))
         
-        if (inherits(max_slope, "numeric")) {
-            slope@transitionMatrix@x[slope@transitionMatrix@x == 6 * exp(-3.5 * abs(rand + 0.05))] <- 0
+        cf <- function(x) {
+            (6 * exp(-3.5 * abs(x[adj] + 0.05)))
         }
         
-        Conductance <- gdistance::geoCorrection(slope)
+    }
+    
+    if (cost_function == "tobler offpath") {
+        
+        cf <- function(x) {
+            (6 * exp(-3.5 * abs(x[adj] + 0.05))) * 0.6
+        }
+        
+    }
+    
+    if (cost_function == "irmischer-clarke male") {
+        
+        cf <- function(x) {
+            (0.11 + exp(-(abs(x[adj]) * 100 + 5)^2/(2 * 30^2)))
+        }
+        
+    }
+    
+    if (cost_function == "irmischer-clarke offpath male") {
+        
+        cf <- function(x) {
+            (0.11 + 0.67 * exp(-(abs(x[adj]) * 100 + 2)^2/(2 * 30^2)))
+        }
+        
+    }
+    
+    if (cost_function == "irmischer-clarke female") {
+        
+        cf <- function(x) {
+            0.95 * (0.11 + exp(-(abs(x[adj]) * 100 + 5)^2/(2 * 30^2)))
+        }
+        
+    }
+    
+    if (cost_function == "irmischer-clarke offpath female") {
+        
+        cf <- function(x) {
+            0.95 * (0.11 + 0.67 * exp(-(abs(x[adj]) * 100 + 2)^2/(2 * 30^2)))
+        }
+        
     }
     
     if (cost_function == "modified tobler") {
-        slope[adj] <- 4.8 * exp(-5.3 * abs((slope[adj] * 0.7) + 0.03))
         
-        if (inherits(max_slope, "numeric")) {
-            slope@transitionMatrix@x[slope@transitionMatrix@x == 4.8 * exp(-5.3 * abs((rand * 0.7) + 0.03))] <- 0
+        cf <- function(x) {
+            (4.8 * exp(-5.3 * abs((x[adj] * 0.7) + 0.03)))
         }
-        Conductance <- gdistance::geoCorrection(slope)
+        
     }
     
     if (cost_function == "wheeled transport") {
-        slope[adj] <- 1/(1 + abs(slope[adj] * 100/crit_slope)^2)
-        if (inherits(max_slope, "numeric")) {
-            slope@transitionMatrix@x[slope@transitionMatrix@x == 1/(1 + abs(rand * 100/crit_slope)^2)] <- 0
+        
+        cf <- function(x) {
+            (1/(1 + abs(x[adj] * 100/crit_slope)^2))
         }
-        Conductance <- gdistance::geoCorrection(slope)
+        
     }
     
     if (cost_function == "herzog") {
-        slope[adj] <- 1/((1337.8 * slope[adj]^6) + (278.19 * slope[adj]^5) - (517.39 * slope[adj]^4) - (78.199 * slope[adj]^3) + (93.419 * slope[adj]^2) + 
-            (19.825 * slope[adj]) + 1.64)
-        if (inherits(max_slope, "numeric")) {
-            slope@transitionMatrix@x[slope@transitionMatrix@x == 1/((((1337.8 * rand^6) + (278.19 * rand^5) - (517.39 * rand^4) - (78.199 * rand^3) + (93.419 * 
-                rand^2) + (19.825 * rand) + 1.64)))] <- 0
+        
+        cf <- function(x) {
+            (1/((1337.8 * x[adj]^6) + (278.19 * x[adj]^5) - (517.39 * x[adj]^4) - (78.199 * x[adj]^3) + (93.419 * x[adj]^2) + (19.825 * x[adj]) + 
+                1.64))
         }
+        
+    }
+    
+    if (cost_function == "llobera-sluckin") {
+        
+        cf <- function(x) {
+            (1/(2.635 + (17.37 * abs(x[adj])) + (42.37 * abs(x[adj])^2) - (21.43 * abs(x[adj])^3) + (14.93 * abs(x[adj])^4)))
+        }
+        
+    }
+    
+    if (cost_function %in% cfs[!cfs == "all"]) {
+        
+        
+        slope[adj] <- cf(slope)
+        
+        if (inherits(max_slope, "numeric")) {
+            slope@transitionMatrix@x[slope@transitionMatrix@x == cf(rand)] <- 0
+        }
+        
         Conductance <- gdistance::geoCorrection(slope)
     }
     
-    if (cost_function == "all") {
+    if (cost_function %in% cfs[cfs == "all"]) {
         
-        slope_stack <- raster::stack(slope, slope, slope, slope)
+        slope_stack <- list(slope, slope, slope, slope, slope, slope, slope, slope, slope, slope)
         
-        slope_stack[[1]][adj] <- 6 * exp(-3.5 * abs(slope[adj] + 0.05))
-        slope_stack[[1]]@transitionMatrix@x[slope@transitionMatrix@x == 6 * exp(-3.5 * abs(rand + 0.05))] <- 0
+        cf1 <- function(x) {
+            (6 * exp(-3.5 * abs(x[adj] + 0.05)))
+        }
+        cf2 <- function(x) {
+            (6 * exp(-3.5 * abs(x[adj] + 0.05))) * 0.6
+        }
         
-        slope_stack[[2]][adj] <- 4.8 * exp(-5.3 * abs((slope[adj] * 0.7) + 0.03))
-        slope_stack[[2]]@transitionMatrix@x[slope@transitionMatrix@x == 4.8 * exp(-5.3 * abs((rand * 0.7) + 0.03))] <- 0
+        cf3 <- function(x) {
+            (0.11 + exp(-(abs(x[adj]) * 100 + 5)^2/(2 * 30^2)))
+        }
         
-        slope_stack[[3]][adj] <- 1/(1 + abs(slope[adj] * 100/crit_slope)^2)
-        slope_stack[[3]]@transitionMatrix@x[slope@transitionMatrix@x == 1/(1 + abs(rand * 100/crit_slope)^2)] <- 0
+        cf4 <- function(x) {
+            (0.11 + 0.67 * exp(-(abs(x[adj]) * 100 + 2)^2/(2 * 30^2)))
+        }
         
-        slope_stack[[4]][adj] <- 1/((((1337.8 * slope[adj]^6) + (278.19 * slope[adj]^5) - (517.39 * slope[adj]^4) - (78.199 * slope[adj]^3) + (93.419 * slope[adj]^2) + 
-            (19.825 * slope[adj]) + 1.64)))
-        slope_stack[[4]]@transitionMatrix@x[slope@transitionMatrix@x == 1/((((1337.8 * rand^6) + (278.19 * rand^5) - (517.39 * rand^4) - (78.199 * rand^3) + 
-            (93.419 * rand^2) + (19.825 * rand) + 1.64)))] <- 0
+        cf5 <- function(x) {
+            0.95 * (0.11 + exp(-(abs(x[adj]) * 100 + 5)^2/(2 * 30^2)))
+        }
         
-        Conductance <- raster::stack(gdistance::geoCorrection(slope_stack[[1]]), gdistance::geoCorrection(slope_stack[[2]]), gdistance::geoCorrection(slope_stack[[3]]), 
-            gdistance::geoCorrection(slope_stack[[4]]))
+        cf6 <- function(x) {
+            0.95 * (0.11 + 0.67 * exp(-(abs(x[adj]) * 100 + 2)^2/(2 * 30^2)))
+        }
+        
+        cf7 <- function(x) {
+            (4.8 * exp(-5.3 * abs((x[adj] * 0.7) + 0.03)))
+        }
+        cf8 <- function(x) {
+            (1/(1 + abs(x[adj] * 100/crit_slope)^2))
+        }
+        cf9 <- function(x) {
+            (1/((1337.8 * x[adj]^6) + (278.19 * x[adj]^5) - (517.39 * x[adj]^4) - (78.199 * x[adj]^3) + (93.419 * x[adj]^2) + (19.825 * x[adj]) + 
+                1.64))
+        }
+        cf10 <- function(x) {
+            (1/(2.635 + (17.37 * abs(x[adj])) + (42.37 * abs(x[adj])^2) - (21.43 * abs(x[adj])^3) + (14.93 * abs(x[adj])^4)))
+        }
+        
+        slope_stack[[1]][adj] <- cf1(slope)
+        slope_stack[[2]][adj] <- cf2(slope)
+        slope_stack[[3]][adj] <- cf3(slope)
+        slope_stack[[4]][adj] <- cf4(slope)
+        slope_stack[[5]][adj] <- cf5(slope)
+        slope_stack[[6]][adj] <- cf6(slope)
+        slope_stack[[7]][adj] <- cf7(slope)
+        slope_stack[[8]][adj] <- cf8(slope)
+        slope_stack[[9]][adj] <- cf9(slope)
+        slope_stack[[10]][adj] <- cf10(slope)
+        
+        if (inherits(max_slope, "numeric")) {
+            slope_stack[[1]]@transitionMatrix@x[slope_stack[[1]]@transitionMatrix@x == cf1(rand)] <- 0
+            slope_stack[[2]]@transitionMatrix@x[slope_stack[[2]]@transitionMatrix@x == cf2(rand)] <- 0
+            slope_stack[[3]]@transitionMatrix@x[slope_stack[[3]]@transitionMatrix@x == cf3(rand)] <- 0
+            slope_stack[[4]]@transitionMatrix@x[slope_stack[[4]]@transitionMatrix@x == cf4(rand)] <- 0
+            slope_stack[[5]]@transitionMatrix@x[slope_stack[[5]]@transitionMatrix@x == cf5(rand)] <- 0
+            slope_stack[[6]]@transitionMatrix@x[slope_stack[[6]]@transitionMatrix@x == cf6(rand)] <- 0
+            slope_stack[[7]]@transitionMatrix@x[slope_stack[[6]]@transitionMatrix@x == cf7(rand)] <- 0
+            slope_stack[[8]]@transitionMatrix@x[slope_stack[[8]]@transitionMatrix@x == cf8(rand)] <- 0
+            slope_stack[[9]]@transitionMatrix@x[slope_stack[[9]]@transitionMatrix@x == cf9(rand)] <- 0
+            slope_stack[[10]]@transitionMatrix@x[slope_stack[[10]]@transitionMatrix@x == cf10(rand)] <- 0
+        }
+        
+        
+        Conductance <- lapply(slope_stack, function(x) {
+            gdistance::geoCorrection(x)
+        })
+        
+        names(Conductance) <- cfs[!cfs == "all"]
+        
     }
     
     return(Conductance)