+++ inst/doc/leastcostpath-1.R
@@ -1,4 +1,4 @@
-## ----libraries, echo = TRUE, message= FALSE, warning= FALSE--------------
+## ----libraries, echo = TRUE, message= FALSE, warning= FALSE-------------------
 library(rgdal)
 library(rgeos)
 library(sp)
@@ -6,7 +6,7 @@ library(raster)
 library(gdistance)
 library(leastcostpath)
 
-## ----raster, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
+## ----raster, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE-----
 r <- raster::raster(system.file('external/maungawhau.grd', package = 'gdistance'))
 
 plot(r)
@@ -14,7 +14,7 @@ plot(r)
 ## ----slope_cs, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
 cs <- create_slope_cs(dem = r, cost_function = 'tobler', neighbours = 16)
 
-## ----lcp, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
+## ----lcp, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE--------
 
 loc1 = cbind(2667876, 6479424)
 loc1 = sp::SpatialPoints(loc1)
@@ -32,7 +32,7 @@ plot(lcp[[2]], add = T, col = "blue")
 cs <- create_slope_cs(dem = r, cost_function = 'tobler', neighbours = 16) %>%
   "*" (create_traversal_cs(dem = r, neighbours = 16))
 
-## ----lcp_2, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
+## ----lcp_2, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE------
 
 loc1 = cbind(2667876, 6479424)
 loc1 = sp::SpatialPoints(loc1)
@@ -56,7 +56,7 @@ cs <- create_slope_cs(dem = r, cost_function = 'tobler', neighbours = 16) %>%
   "*" (create_traversal_cs(dem = r, neighbours = 16)) %>%
   "*" (create_feature_cs(raster = r, locations = feature_loc, x))
 
-## ----lcp_3, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
+## ----lcp_3, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE------
 
 loc1 = cbind(2667876, 6479424)
 loc1 = sp::SpatialPoints(loc1)
@@ -81,7 +81,7 @@ cs <- create_slope_cs(dem = r, cost_function = 'tobler', neighbours = 16) %>%
   "*" (create_traversal_cs(dem = r, neighbours = 16)) %>%
   "*" (create_feature_cs(raster = r, locations = feature_loc, x))
 
-## ----cc, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
+## ----cc, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE---------
 
 loc1 = cbind(2667876, 6479424)
 loc1 = sp::SpatialPoints(loc1)
@@ -93,7 +93,7 @@ cc <- create_cost_corridor(cs, loc1, loc2)
 
 plot(cc)
 
-## ----fete, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
+## ----fete, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE-------
 locs <- sp::spsample(as(r, 'SpatialPolygons'),n=25,'regular')
 
 lcp_network <- create_slope_cs(dem = r, cost_function = 'tobler', neighbours = 16) %>%
@@ -104,12 +104,12 @@ plot(r)
 plot(locs, add = T)
 plot(lcp_network, add = T, col = "red")
 
-## ----lcp_d, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
+## ----lcp_d, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE------
 lcp_network_density <- create_lcp_density(lcps = lcp_network, raster = r, rescale = TRUE)
 
 plot(lcp_network_density)
 
-## ----lcp_kd, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
+## ----lcp_kd, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE-----
 fwModel <- raster::focalWeight(lcp_network_density, max(res(r)) * 2, type='circle')
 
 fwModel[fwModel>0] <- 1
@@ -118,7 +118,7 @@ lcp_network_kd <- raster::focal(lcp_network_density, w=fwModel ,fun=sum , na.rm=
 
 plot(lcp_network_kd)
 
-## ----ccp, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
+## ----ccp, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE--------
 locs = cbind(2667652, 6478997)
 locs = sp::SpatialPoints(locs)
 
@@ -137,6 +137,17 @@ locs = sp::SpatialPoints(locs)
 lcp_network <- create_slope_cs(dem = r, cost_function = 'tobler', neighbours = 16) %>%
   "*" (create_traversal_cs(dem = r, neighbours = 16)) %>%
   create_banded_lcps(cost_surface = ., location = locs, min_distance = 5, max_distance = 250, radial_points = 50, cost_distance = FALSE, parallel = FALSE)
+
+plot(r)
+plot(locs, add = T)
+plot(lcp_network, add = T, col = "red")
+
+## ----lcp_network, echo = TRUE,  fig.height = 6, fig.width = 6, warning = FALSE----
+locs <- sp::spsample(as(r, 'SpatialPolygons'),n=5,'regular')
+
+lcp_network <- create_slope_cs(dem = r, cost_function = 'tobler', neighbours = 16) %>%
+  "*" (create_traversal_cs(dem = r, neighbours = 16)) %>%
+  create_lcp_network(., locations = locs, nb_matrix = cbind(c(1, 4, 2, 1), c(2, 2, 4, 3)), cost_distance = FALSE, parallel = FALSE)
 
 plot(r)
 plot(locs, add = T)