+++ NEWS.md
@@ -1,3 +1,15 @@
+# 1.2.3 (28 April 2020)
+* Added create_lcp_network vignette
+* Fixed create_lcp_density issue where raster values of 1 would become 0 when rescale is TRUE
+* Addded Tobler's offpath cost function
+* Added irmischer-clarke male/female cost function
+* Added irmischer-clarke offpath male/female cost function
+* Added llobera-sluckin cost function
+* Refactored create_slope_cs for easier scalability
+
+# 1.2.2 (01 April 2020)
+* Implemented create_lcp_network
+
 # 1.2.1 (18 March 2020)
 * Maximum slope traversable argument added to create_slope_cs function